[WIP-ish] Implement initial gear rotation generator, refactor unit tests to be worse
- MC-TextureGen now generates textures for every angle of the gear rotation animation! The code is very unreadable right now, and needs to be refactored.
- The TextureGenerator class now has some math utility methods to help with the gear rotation generator. These utility methods should produce identical results to Minecraft's, because they're both based on Riven's code.
- The unit tests are now worse. I've implemented a messy way of signalling that something went wrong with a TextureGenerator, but this needs to be reworked.

I've had this code nearly finished for some weeks now, and I've decided that I might as well commit it as-is. I've been moving into a new house, and haven't had much time to work on this project, so this code is not really up to standard. I'll probably refactor it a bit before a new release.